WIP: ignore sync errors for duplicate heights
Ignore sync errors when a different block has already been verified at
this height by the checkpoint verifier.

We might not want this fix, because we want to restart the sync when we
are following a side-chain.

There is a design tradeoff here - if we reset for side-chains, bad peers
can force us to stall. But if we don't reset, bad peers can force us to
follow a side-chain.

A peer reputation system might help with this tradeoff.